use @ton/* package and refactor code
