+++ package.json
@@ -9,30 +9,27 @@
     "d2": "ts-node ./sources/2_TransferURL.ts"
   },
   "dependencies": {
+    "@aws-crypto/sha256-js": "^5.2.0",
+    "@tact-lang/compiler": "^1.1.5",
+    "@tact-lang/emulator": "^4.2.3",
     "@ton-community/sandbox": "^0.11.0",
     "@ton-community/test-utils": "^0.3.0",
-    "ton-emulator": "^2.1.1",
-    
-    "ton-core": "^0.51.0",
-    "ton": "^13.6.0",
-
-    "dotenv": "^16.3.1",
-    "@aws-crypto/sha256-js": "^4.0.0",
-    "@tact-lang/compiler": "^1.1.3",
-    "@tact-lang/emulator": "^3.2.2",
-    "@types/jest": "^29.5.3",
-    "@types/node": "^20.2.5",
-    "@types/qs": "^6.9.7",
+    "@ton/core": "^0.53.0",
+    "@ton/crypto": "^3.2.0",
+    "@ton/ton": "^13.9.0",
+    "@types/jest": "^29.5.11",
+    "@types/node": "^20.11.5",
+    "@types/qs": "^6.9.11",
     "base64url": "^3.0.1",
-    "enquirer": "^2.3.6",
-    "jest": "^29.6.2",
-    "open": "^8.4.0",
+    "dotenv": "^16.3.2",
+    "enquirer": "^2.4.1",
+    "jest": "^29.7.0",
+    "open": "^10.0.3",
     "prando": "^6.0.1",
-    "prettier": "^2.8.8",
-    "qs": "^6.11.0",
-    "ton-crypto": "^3.2.0",
-    "ts-jest": "^29.1.1",
-    "ts-node": "^10.9.1",
-    "typescript": "^5.1.6"
+    "prettier": "^3.2.4",
+    "qs": "^6.11.2",
+    "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.3.3"
   }
 }
