+++ sources/contract.config.ts
@@ -1,7 +1,7 @@
 export const configJettonParams = {
-    name: "Tact Token Name2",
+    name: "Tact Token Name213",
     description: "This is description of Test Jetton Token in Tact-lang",
-    symbol: "TTN2",
+    symbol: "TTN213",
     image: "https://avatars.githubusercontent.com/u/104382459?s=200&v=4",
 };
