+++ sources/contract.read.ts
@@ -1,30 +1,18 @@
-import {
-    WalletContractV4,
-    beginCell,
-    Address,
-    contractAddress,
-    ContractProvider,
-    TonClient4,
-    TonClient,
-    fromNano,
-    toNano,
-    Cell,
-    BitString,
-    Slice,
-} from "ton";
+
 import { printSeparator } from "./utils/print";
 
 // Contract Abi //
 import { buildOnchainMetadata } from "./utils/jetton-helpers";
-import { mnemonicToPrivateKey } from "ton-crypto";
 
-import { SampleJetton, loadTokenTransfer } from "./output/SampleJetton_SampleJetton";
-import { JettonDefaultWallet } from "./output/SampleJetton_JettonDefaultWallet";
+import {contractAddress, toNano, TonClient4, WalletContractV4} from "@ton/ton";
+import {mnemonicToPrivateKey} from "@ton/crypto";
+import {_ENDPOINT_MAINNET, _ENDPOINT_TESTNET, _IS_TEST_ENV} from "./utils/static";
+import {JettonMasterContract} from "./output/JettonTact_JettonMasterContract";
+import {JettonDefaultWallet} from "./output/JettonTact_JettonDefaultWallet";
 
 (async () => {
-    // //create client for testnet sandboxv4 API - alternative endpoint
     // const client = new TonClient4({
-    //     endpoint: "https://sandbox-v4.tonhubapi.com",
+    //     endpoint: _IS_TEST_ENV ? _ENDPOINT_TESTNET : _ENDPOINT_MAINNET
     // });
     // const mnemonics = "YOUR OWN mnemonics";
     // let keyPair = await mnemonicToPrivateKey(mnemonics.split(" "));
@@ -43,10 +31,10 @@ import { JettonDefaultWallet } from "./output/SampleJetton_JettonDefaultWallet";
     // let max_supply = toNano(1234567666666689011); // Set the specific total supply in nano
     // // Create content Cell
     // let content = buildOnchainMetadata(jettonParams);
-    // let init = await SampleJetton.init(deploy_wallet_contract.address, content, max_supply);
+    // let init = await JettonMasterContract.init(deploy_wallet_contract.address, content, max_supply);
     // let jetton_minter_contract_address = contractAddress(workchain, init);
     // console.log("Jetton Master: " + jetton_minter_contract_address);
-    // let contract_ddd = await client.open(SampleJetton.fromAddress(jetton_minter_contract_address));
+    // let contract_ddd = await client.open(JettonMasterContract.fromAddress(jetton_minter_contract_address));
     // let jetton_wallet = await contract_ddd.getGetWalletAddress(deploy_wallet_contract.address);
     // let contract_dataFormat = JettonDefaultWallet.fromAddress(jetton_wallet);
     // let contract = client.open(contract_dataFormat);
@@ -55,7 +43,7 @@ import { JettonDefaultWallet } from "./output/SampleJetton_JettonDefaultWallet";
     // let owner_of_wallet = await (await contract.getGetWalletData()).owner;
     // console.log("JettonWallet Balance: " + jettonWalletBalance);
     // console.log("JettonWallet Owner: \n" + owner_of_wallet);
-    // TODO:
+    // // TODO:
     // // loadOwnershipAssigned => msg.forwardload
     // let aa = loadTransferEvent(src.asSlice());
     // console.log("Mint MemberID: " + aa.item_index + ", by " + aa.minter);
