+++ sources/master.tact
@@ -5,7 +5,7 @@ message Mint {
     receiver: Address;
 }
 
-contract SampleJetton with Jetton {
+contract JettonMasterContract with Jetton {
     total_supply: Int as coins;
     owner: Address;
     content: Cell;
@@ -33,7 +33,7 @@ contract SampleJetton with Jetton {
         let ctx: Context = context();
         require(self.mintable, "Not mintable");
         require(self.total_supply + 100 <= self.max_supply, "Max supply exceeded");
-        self.mint(ctx.sender, 100, self.owner); // 🔴 
+        self.mint(ctx.sender, 100, self.owner); // 🔴
     }
 
     receive("Owner: MintClose") {
@@ -41,4 +41,4 @@ contract SampleJetton with Jetton {
         require(ctx.sender == self.owner, "Not owner");
         self.mintable = false;
     }
-} 
+}
