+++ sources/output/JettonTact_JettonMasterContract.code.rev.fif
@@ -133,7 +133,7 @@ PROGRAM{
     s0 s1 XCHG
   }>
   get_abi_ipfs PROC:<{
-    x{697066733A2F2F516D58416837326737786951393753593550616F4534526D645A4D59634642576A684B63726A6F6371416B67514782_} PUSHSLICE
+    x{697066733A2F2F516D664C514E644A66354B5961784E437950726E553575745A32367653677146514C3938676E4B584D594259753182_} PUSHSLICE
   }>
   ?fun_ref_10c319ce0484d846 PROCREF:<{
     s3 PUSH
