+++ sources/output/JettonTact_JettonMasterContract.storage.fc
@@ -202,10 +202,10 @@ builder $JettonDefaultWallet$init$_store(builder build_0, (slice, slice) v) inli
 }
 
 ;;
-;; Type: SampleJetton
+;; Type: JettonMasterContract
 ;;
 
-builder $SampleJetton$_store(builder build_0, (int, slice, cell, int, int) v) inline {
+builder $JettonMasterContract$_store(builder build_0, (int, slice, cell, int, int) v) inline {
     var (v'total_supply, v'owner, v'content, v'mintable, v'max_supply) = v;
     build_0 = build_0.store_coins(v'total_supply);
     build_0 = __tact_store_address(build_0, v'owner);
@@ -215,7 +215,7 @@ builder $SampleJetton$_store(builder build_0, (int, slice, cell, int, int) v) in
     return build_0;
 }
 
-(slice, ((int, slice, cell, int, int))) $SampleJetton$_load(slice sc_0) inline {
+(slice, ((int, slice, cell, int, int))) $JettonMasterContract$_load(slice sc_0) inline {
     var v'total_supply = sc_0~load_coins();
     var v'owner = sc_0~__tact_load_address();
     var v'content = sc_0~load_ref();
@@ -224,32 +224,32 @@ builder $SampleJetton$_store(builder build_0, (int, slice, cell, int, int) v) in
     return (sc_0, (v'total_supply, v'owner, v'content, v'mintable, v'max_supply));
 }
 
-(slice, ((slice, cell, int))) $SampleJetton$init$_load(slice sc_0) inline {
+(slice, ((slice, cell, int))) $JettonMasterContract$init$_load(slice sc_0) inline {
     var v'owner = sc_0~__tact_load_address();
     var v'content = sc_0~load_ref();
     var v'max_supply = sc_0~load_int(257);
     return (sc_0, (v'owner, v'content, v'max_supply));
 }
 
-(int, slice, cell, int, int) $SampleJetton$_contract_load() impure inline_ref {
+(int, slice, cell, int, int) $JettonMasterContract$_contract_load() impure inline_ref {
     slice $sc = get_data().begin_parse();
     __tact_context_sys = $sc~load_ref();
     int $loaded = $sc~load_int(1);
     if ($loaded) {
-        return $sc~$SampleJetton$_load();
+        return $sc~$JettonMasterContract$_load();
     } else {
         ;; Allow only workchain deployments
         throw_unless(137, my_address().preload_uint(11) == 1024);
-        (slice owner, cell content, int max_supply) = $sc~$SampleJetton$init$_load();
+        (slice owner, cell content, int max_supply) = $sc~$JettonMasterContract$init$_load();
         $sc.end_parse();
-        return $SampleJetton$_contract_init(owner, content, max_supply);
+        return $JettonMasterContract$_contract_init(owner, content, max_supply);
     }
 }
 
-() $SampleJetton$_contract_store((int, slice, cell, int, int) v) impure inline {
+() $JettonMasterContract$_contract_store((int, slice, cell, int, int) v) impure inline {
     builder b = begin_cell();
     b = b.store_ref(__tact_context_sys);
     b = b.store_int(true, 1);
-    b = $SampleJetton$_store(b, v);
+    b = $JettonMasterContract$_store(b, v);
     set_data(b.end_cell());
 }
