--- sources/output/SampleJetton_SampleJetton.constants.fc
@@ -1,2 +1,0 @@
-;; Comment "Approve1234"
-cell __gen_cell_comment_87d24fc243dec9b8d5076503e6a0444f7d6d6bcdb25a538fcf14e38f4bd6e075() asm "B{b5ee9c7241010101001100001e00000000417070726f76653132333426c1b169} B>boc PUSHREF";
